agent/requirements.txt
uagents>=0.22.0
pydantic>=2.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.19.0
python-dotenv>=1.0.0
httpx>=0.25.0
asyncio-mqtt>=0.16.0
